removed function in mongodb
